SearchView :: Separa componentes em arquivos proprios
