+++ src/components/FavoritesView/index.js
@@ -1,12 +1,9 @@
-import React, { Fragment, useEffect } from 'react'
+import React, { useEffect } from 'react'
 import styled from 'styled-components'
-import { Row, Cell } from 'griding'
 import { useFavorites } from 'utils/favorites'
 import Container from 'components/Container'
-import FetchCard from 'components/Card/FetchCard'
-import InfoScreen from 'components/InfoScreen'
-import Text from 'components/Text'
-import { HeartBreak } from 'components/Icon'
+import HasFavorites from './has-favorites'
+import NoFavorites from './no-favorites'
 
 const Wrapper = styled.div`
 	flex: 1
@@ -16,42 +13,6 @@ const Wrapper = styled.div`
 `
 
 const getTitle = length => `${!length ? 'no favorites' : `${length} ${length > 1 ? `favorites`: `favorite`}`}`
-
-const NoFavorites = ({length}) => {
-	if (!length) {
-		return (
-			<InfoScreen
-				emoji={<HeartBreak size={96} style={{margin: '1rem'}}/>}
-				title='You still have no favorites'
-				description='you can add movies to your favorites clicking on the ♥ icon'
-			/>
-		)
-	}
-
-	return null;
-}
-
-const HasFavorites = ({favorites}) => {
-	if (favorites.length) {
-		return (
-			<Fragment>
-				<Text weight={600} xs={2} sm={3} md={4} xg={5}>{getTitle(favorites.length)}</Text>
-				<Row vertical-gutter style={{marginTop: '2rem', marginBottom: '2rem'}}>
-					{favorites.map(favorite => (
-						<Cell key={favorite} xs={6} sm={4} md={3} xg={2}>
-							<FetchCard
-								kindURL={favorite?.split('/')[0]}
-								id={favorite?.split('/')[1]}
-							/>
-						</Cell>
-					))}
-				</Row>
-			</Fragment>
-		)
-	}
-
-	return null;
-}
 
 const SearchView = () => {
 	const [favoritesSet] = useFavorites()
